+++ src/engine/Scene.ts
@@ -721,7 +721,7 @@ export class Scene implements StorageLoaderHandler {
 	 * @param e 子エンティティとして追加するエンティティ
 	 */
 	append(e: E): void {
-		this.insertBefore(e);
+		this.insertBefore(e, undefined);
 	}
 
 	/**
@@ -733,7 +733,7 @@ export class Scene implements StorageLoaderHandler {
 	 * @param e 子エンティティとして追加するエンティティ
 	 * @param target 挿入位置にある子エンティティ
 	 */
-	insertBefore(e: E, target?: E): void {
+	insertBefore(e: E, target: E | undefined): void {
 		if (e.parent) e.remove();
 
 		e.parent = this;