+++ src/pdi-common-impls/TextAsset.ts
@@ -13,8 +13,8 @@ export abstract class TextAsset extends Asset implements TextAssetLike {
 	type: "text" = "text";
 	data: string;
 
-	constructor(id: string, path: string) {
-		super(id, path);
+	constructor(id: string, assetPath: string) {
+		super(id, assetPath);
 		this.data = undefined!;
 	}